impls/union.rs
