+++ src/impls.rs
@@ -1,3 +1,4 @@
 mod add;
 mod index;
 mod intersection;
+mod union;
